Rename file and usage
